+++ docs/CHANGELOG.md
@@ -20,6 +20,7 @@ and this project (aims to) adhere to [Semantic Versioning](https://semver.org/sp
 ### Changed
 - updated Avalonia to 12.1.1 and Avalonia WebView to 12.1.0
 - expand and unify item context menus
+- stop music playback when encountering defect music files
 
 ### Fixed
 - wayland starts in fullscreen